makefile

+++ TRABALHO_ULTIMA VERSAO/operating_system.cpp
@@ -9,7 +9,7 @@ Operating_system::Operating_system() {
 }
 
 Operating_system::~Operating_system() {
-    for (int i = 0; i < processes_queue.size(); i++) {
+    for (int i = 0; i < static_cast<int>(processes_queue.size()); i++) {
         Process* p = processes_queue[i];
         delete p;
     }
@@ -98,7 +98,7 @@ void Operating_system::run_scheduler() {
 // Verificação de término
 bool Operating_system::verify_finished_scheduling(){
     int allFinished = true;
-    for (int i = 0; i < processes_queue.size(); i++) {
+    for (int i = 0; i < static_cast<int>(processes_queue.size()); i++) {
         if (processes_queue[i]->get_state() != "Finished") {
             allFinished = false;
             return allFinished;
@@ -109,7 +109,7 @@ bool Operating_system::verify_finished_scheduling(){
 
 // Imprimir na tela os atributos dos processos
 void Operating_system::print_processes_queue(){
-    for (int i = 0; i < processes_queue.size(); i++) {
+    for (int i = 0; i < static_cast<int>(processes_queue.size()); i++) {
         cout << "ID: " << processes_queue[i]->get_id() << endl;
         cout << "DURATION: " << processes_queue[i]->get_duration() << endl;
         cout << "CREATION TIME: " << processes_queue[i]->get_creation_time() << endl;
